refactor(primitves): Consistent naming of primitves (files and cli)
