+++ src/gallia/commands/__init__.py
@@ -11,15 +11,15 @@ from gallia.commands.primitive.generic.pdu import GenericPDUPrimitive
 from gallia.commands.primitive.uds.dtc import DTCPrimitive
 from gallia.commands.primitive.uds.ecu_reset import ECUResetPrimitive
 from gallia.commands.primitive.uds.iocbi import IOCBIPrimitive
+from gallia.commands.primitive.uds.pdu import SendPDUPrimitive
 from gallia.commands.primitive.uds.ping import PingPrimitive
-from gallia.commands.primitive.uds.read_by_identifier import ReadByIdentifierPrimitive
+from gallia.commands.primitive.uds.rdbi import ReadByIdentifierPrimitive
 from gallia.commands.primitive.uds.rmba import RMBAPrimitive
 from gallia.commands.primitive.uds.rtcl import RTCLPrimitive
-from gallia.commands.primitive.uds.send_pdu import SendPDUPrimitive
-from gallia.commands.primitive.uds.test_xcp import SimpleTestXCP
 from gallia.commands.primitive.uds.vin import VINPrimitive
+from gallia.commands.primitive.uds.wdbi import WriteByIdentifierPrimitive
 from gallia.commands.primitive.uds.wmba import WMBAPrimitive
-from gallia.commands.primitive.uds.write_by_identifier import WriteByIdentifierPrimitive
+from gallia.commands.primitive.uds.xcp import SimpleTestXCP
 from gallia.commands.scan.uds.identifiers import ScanIdentifiers
 from gallia.commands.scan.uds.memory import MemoryFunctionsScanner
 from gallia.commands.scan.uds.reset import ResetScanner
